+++ packages/mobile/src/content-suggest/ContentSuggest.tsx
@@ -1,0 +1,170 @@
+import React, {Fragment} from 'react'
+
+import {
+  ModalSuggestControl,
+  Image,
+  MenuControl,
+  Icon,
+  ContentInput,
+  Pos,
+  Box,
+  Card,
+  Spacer,
+  styled,
+} from '@qiwi/pijma-core'
+
+import {Paragraph} from '../typography'
+import {Link} from '../link'
+import {MenuItem} from '../menu'
+import {InputModal} from '../input-modal'
+
+import ContentSuggestProps from './ContentSuggestProps'
+import ContentSuggestOptionModel from './ContentSuggestOptionModel'
+
+const CardItem = styled(Card)().withComponent(MenuItem)
+
+export const ContentSuggest = <V extends {}>({
+  equals = (a: V, b: V) => a === b,
+  ...props
+}: ContentSuggestProps<ContentSuggestOptionModel<V>, V>) => (
+  <ModalSuggestControl<V>
+    value={props.value}
+    suggest={props.suggest}
+    items={props.items}
+    total={props.total}
+    empty={props.empty}
+    equals={equals}
+    onRequest={props.onRequest}
+    onChange={props.onChange}
+    onBlur={props.onBlur}
+    onFocus={props.onFocus}
+    onCancel={props.onCancel}
+    onSubmit={props.onSubmit}
+    children={(renderProps) => (
+      <Pos type="relative">
+        <Box
+          width={1}
+          onMouseEnter={renderProps.onMouseEnter}
+          onMouseLeave={renderProps.onMouseLeave}
+        >
+          <ContentInput
+            value={props.suggest || ''}
+            type="search"
+            error={false}
+            tabIndex={props.tabIndex}
+            autoComplete={props.autoComplete}
+            autoFocus={props.autoFocus}
+            placeholder={props.placeholder}
+            maxLength={props.maxLength}
+            pr={14}
+            focused={renderProps.focused}
+            hovered={renderProps.hovered}
+            onChange={renderProps.onRequest}
+            onFocus={renderProps.onFocus}
+          />
+          <Pos type="absolute" right={4} top={3} onClick={renderProps.onShow}>
+            <Icon name="search" color="#666"/>
+          </Pos>
+        </Box>
+        <MenuControl
+          count={props.items.length}
+          selected={renderProps.selected}
+          onSelect={renderProps.onSelect}
+          onKeyDown={renderProps.onKeyDown}
+          children={(menuRenderProps) => (
+            <InputModal
+              value={props.suggest || ''}
+              tabIndex={props.tabIndex}
+              autoComplete={props.autoComplete}
+              placeholder={props.placeholder}
+              maxLength={props.maxLength}
+              show={renderProps.show}
+              inputRef={renderProps.inputRef}
+              contentRef={menuRenderProps.containerRef}
+              error={props.error}
+              onChange={renderProps.onRequest}
+              onKeyDown={renderProps.show ? menuRenderProps.onKeyDown : renderProps.onKeyDown}
+              onBlur={renderProps.onModalInputBlur}
+              onSubmit={renderProps.onSearchClick}
+              onShow={renderProps.onShow}
+              onHide={renderProps.onHide}
+              onEscape={renderProps.onEscape}
+              onBack={renderProps.onBack}
+            >
+              {props.loading ? (
+                Array(4).fill(1).map((_item, key) => (
+                  <CardItem key={key} icon={true} stub text="stub" notes="stub"/>
+                ))
+              ) : (
+                <Spacer size="s">
+                  {menuRenderProps.items.length > 0 ? (
+                    <Fragment>
+                      {menuRenderProps.items.map((item, key) => (
+                        <CardItem
+                          key={key}
+                          ref={item.ref}
+                          onClick={item.onClick}
+                          onMouseEnter={item.onMouseEnter}
+                          cursor="pointer"
+                          text={props.items[key].title}
+                          notes={props.items[key].description}
+                          icon={<Image width={6} height={6} src={props.items[key].logo}/>}
+                          hover={item.focused}
+                          active={item.selected}
+                          focus={item.selected}
+                        />
+                      ))}
+                    </Fragment>
+                  ) : (
+                    null
+                  )}
+                  {props.total && props.items.length > 0 ? (
+                    <Box px={4}>
+                      <Paragraph>
+                        {props.total.text}
+                        {props.total.link ? (
+                          <Fragment>
+                            {' '}
+                            <Link
+                              onClick={renderProps.onTotalClick}
+                              children={props.total.link.text}
+                            />
+                          </Fragment>
+                        ) : (
+                          null
+                        )}
+                      </Paragraph>
+                    </Box>
+                  ) : props.empty && props.items.length === 0 ? (
+                    <Box px={4}>
+                      <Paragraph>
+                        {props.empty.text}
+                        {props.empty.link ? (
+                          <Fragment>
+                            {' '}
+                            <Link
+                              onClick={renderProps.onEmptyClick}
+                              children={props.empty.link.text}
+                            />
+                          </Fragment>
+                        ) : (
+                          null
+                        )}
+                      </Paragraph>
+                    </Box>
+                  ) : (
+                    null
+                  )}
+                </Spacer>
+              )}
+            </InputModal>
+          )}
+        />
+      </Pos>
+    )}
+  />
+)
+
+ContentSuggest.defaultProps = {
+  equals: (a: any, b: any) => a === b,
+}